requirements.txt
streamlit
opencv-python
pandas
numpy
Pillow
